+++ src/GameDisplay.cpp
@@ -11,19 +11,19 @@ String GameDisplay::loadingStr;
 GameDisplay::GameDisplay(sf::RenderWindow* wnd)
     : renderWnd(wnd), error(false)
 {
-	this->wndSizeDefault = renderWnd->getSize();
+	wndSizeDefault = renderWnd->getSize();
 	loadingStr = "Loading GameDisplay...";
 
     cout << "GameDisplay: Creating GameDisplay instance..." << endl;
     instance = this;
 
-	this->fullscreenMode = 0;
-	this->vsync = true;
+	fullscreenMode = 0;
+	vsync = true;
 }
 
 void GameDisplay::clearTextures()
 {
-    this->texturesByName.clear();
+    texturesByName.clear();
 }
 
 void GameDisplay::reload()
@@ -31,7 +31,7 @@ void GameDisplay::reload()
     cout << "GameDisplay: Reloading graphics resources..." << endl;
     GameDisplay::loadingStr = "Reloading resources...";
 
-    this->clearTextures();
+    clearTextures();
 
     if(unknownTexture.getSize() == Vector2u(0,0))
     {
@@ -50,18 +50,18 @@ void GameDisplay::reload()
             // %%%%%%%*******
             // %%%%%%%******* 300
 
-            bool isNB = ((i <= 32 && j >= 32) || (i >= 32 && j <= 32));
+            bool isNB = ((i <= 32 && j > 32) || (i > 32 && j <= 32));
             imgUnknownTexture.setPixel(i, j, isNB ? Color::Red : Color::Green);
         }
-        this->unknownTexture.loadFromImage(imgUnknownTexture);
-        this->unknownTexture.setRepeated(true);
+        unknownTexture.loadFromImage(imgUnknownTexture);
+        unknownTexture.setRepeated(true);
     }
 
 	for(int i = 0; i < Car::COUNT; i++)
 	{
 	    CarType* type = Game::instance->findCarTypeByID(Car::TypeId(i));
 		if(type != NULL)
-            this->addTexture("car/" + type->getTextureName());
+            addTexture("car/" + type->getTextureName());
         else
             cout << "GameDisplay: unknown car type " << i << ", cannot load textures." << endl;
 	}
@@ -70,21 +70,21 @@ void GameDisplay::reload()
     {
         for(auto ld : Game::instance->levelRegistry)
         {
-            this->addTexture("bg/" + ld.second->getTextureName(), true, true);
-            this->addTexture("map/" + ld.second->getTextureName());
+            addTexture("bg/" + ld.second->getTextureName(), true, true);
+            addTexture("map/" + ld.second->getTextureName());
         }
     }
     else
         cout << "GameDisplay: cannot load map textures, no level registered." << endl;
 
-    this->addTexture("stat/coin");
-    this->addTexture("stat/high");
-    this->addTexture("stat/score");
-    this->addTexture("stat/mpl");
-    this->addTexture("stat/points_mpl");
-    this->addTexture("gui/start");
-    this->addTexture("gui/settings");
-    this->addTexture("gui/quit");
+    addTexture("stat/coin");
+    addTexture("stat/high");
+    addTexture("stat/score");
+    addTexture("stat/mpl");
+    addTexture("stat/points_mpl");
+    addTexture("gui/start");
+    addTexture("gui/settings");
+    addTexture("gui/quit");
 
     for(auto it = Game::instance->powerRegistry.begin(); it != Game::instance->powerRegistry.end(); it++)
     {
@@ -101,14 +101,14 @@ void GameDisplay::addTexture(string name, bool repeated, bool smooth)
 	if (!load)
 	{
 	    cout << "GameDisplay: err D00(name='" << name << "')" << endl;
-		this->texturesByName.insert(pair<string, sf::Texture>(name, unknownTexture));
+		texturesByName.insert(pair<string, sf::Texture>(name, unknownTexture));
 		return;
 	}
 	else
 	{
 		tx.setRepeated(repeated);
 		tx.setSmooth(smooth);
-		this->texturesByName.insert(make_pair(name, tx));
+		texturesByName.insert(make_pair(name, tx));
 	}
 }
 
@@ -153,31 +153,31 @@ void GameDisplay::display()
 
     if(!game->isGuiLoaded())
     {
-        this->drawGame();
+        drawGame();
     }
     else
     {
-        this->renderWnd->clear(sf::Color(50, 40, 40));
+        renderWnd->clear(sf::Color(50, 40, 40));
     }
 
-    this->drawGui();
+    drawGui();
 
 	// Tick and draw splash
 
-	if (this->splashSet)
-		this->drawSplash(this->splashText);
+	if (splashSet)
+		drawSplash(splashText);
 
-	if (this->splashTick > 0)
-		this->splashTick--;
+	if (splashTick > 0)
+		splashTick--;
 	else
-		this->splashSet = false;
+		splashSet = false;
 
-    this->renderWnd->display();
+    renderWnd->display();
 }
 
 sf::Text GameDisplay::drawString(String tx, int height, sf::Vector2f pos, sf::Text::Style style)
 {
-    sf::Text text = sf::Text(tx, this->guiFont);
+    sf::Text text = sf::Text(tx, guiFont);
     text.setPosition(pos);
     text.setCharacterSize(height);
     text.setStyle(style);
@@ -187,7 +187,7 @@ sf::Text GameDisplay::drawString(String tx, int height, sf::Vector2f pos, sf::Te
 
 sf::Text GameDisplay::drawCenteredString(String tx, int height, sf::Vector2f pos, sf::Text::Style style)
 {
-	sf::Text text = sf::Text(tx, this->guiFont);
+	sf::Text text = sf::Text(tx, guiFont);
 	text.setCharacterSize(height);
 	text.setStyle(style);
 	text.setPosition(pos.x - text.getLocalBounds().width / 2, pos.y - text.getLocalBounds().height / 2);
@@ -239,19 +239,19 @@ void GameDisplay::drawGame()
 {
     Game* game = Game::instance;
 
-    this->renderWnd->clear(game->getLevelColor());
+    renderWnd->clear(game->getLevelColor());
 
     // BACKGROUND
-    sf::Sprite bg(this->getTexture("bg/" + game->level.getTextureName().toAnsiString()), IntRect(0, 0, this->getSize().x, 250));
+    sf::Sprite bg(getTexture("bg/" + game->level.getTextureName().toAnsiString()), IntRect(0, 0, getSize().x, 250));
 	bg.setOrigin(0.f, 125.f);
 	bg.setPosition(0.f, this->getSize().y / 2.f);
 	bg.setScale(2.f, 2.f);
-    this->renderWnd->draw(bg);
+    renderWnd->draw(bg);
 
     // FOG
     VertexArray arr(Quads);
     Color fogColor = game->getLevelColor();
-    arr.append(Vertex(Vector2f(0,0), fogColor));
+    /*arr.append(Vertex(Vector2f(0,0), fogColor));
     arr.append(Vertex(Vector2f(this->getSize().x,0), fogColor));
     arr.append(Vertex(Vector2f(this->getSize().x,this->getSize().y/2-250), fogColor));
     arr.append(Vertex(Vector2f(0,this->getSize().y/2-250), fogColor));
@@ -259,22 +259,22 @@ void GameDisplay::drawGame()
     arr.append(Vertex(Vector2f(0,this->getSize().y), fogColor));
     arr.append(Vertex(Vector2f(this->getSize().x,this->getSize().y), fogColor));
     arr.append(Vertex(Vector2f(this->getSize().x,this->getSize().y/2+250), fogColor));
-    arr.append(Vertex(Vector2f(0,this->getSize().y/2+250), fogColor));
+    arr.append(Vertex(Vector2f(0,this->getSize().y/2+250), fogColor));*/
 
 
-    arr.append(Vertex(Vector2f(0,this->getSize().y/2-250), fogColor));
-    arr.append(Vertex(Vector2f(this->getSize().x,this->getSize().y/2-250), fogColor));
-    arr.append(Vertex(Vector2f(this->getSize().x,this->getSize().y/2-200), Color::Transparent));
-    arr.append(Vertex(Vector2f(0,this->getSize().y/2-200), Color::Transparent));
+    arr.append(Vertex(Vector2f(0,getSize().y/2-250), fogColor));
+    arr.append(Vertex(Vector2f(getSize().x,getSize().y/2-250), fogColor));
+    arr.append(Vertex(Vector2f(getSize().x,getSize().y/2-200), Color::Transparent));
+    arr.append(Vertex(Vector2f(0,getSize().y/2-200), Color::Transparent));
 
-    arr.append(Vertex(Vector2f(0,this->getSize().y/2+250), fogColor));
-    arr.append(Vertex(Vector2f(this->getSize().x,this->getSize().y/2+250), fogColor));
-    arr.append(Vertex(Vector2f(this->getSize().x,this->getSize().y/2+200), Color::Transparent));
-    arr.append(Vertex(Vector2f(0,this->getSize().y/2+200), Color::Transparent));
-    this->renderWnd->draw(arr);
+    arr.append(Vertex(Vector2f(0,getSize().y/2+250), fogColor));
+    arr.append(Vertex(Vector2f(getSize().x,getSize().y/2+250), fogColor));
+    arr.append(Vertex(Vector2f(getSize().x,getSize().y/2+200), Color::Transparent));
+    arr.append(Vertex(Vector2f(0,getSize().y/2+200), Color::Transparent));
+    renderWnd->draw(arr);
 
     // CARS
-    sf::Sprite car(this->getTexture("car/default"));
+    sf::Sprite car(getTexture("car/default"));
     car.setOrigin(50.f, 20.f);
     car.setScale(2.0f, 2.0f);
 
@@ -288,7 +288,7 @@ void GameDisplay::drawGame()
         Car& carobj = *game->cars[i];
 		int animFrame = game->mainTickCount/carobj.frameLength % carobj.animSize;
 
-		// health bars
+		// health bars...
         rect1.setSize(sf::Vector2f((carobj.health / carobj.maxHealth) * 30.f, 3.f));
         rect2.setSize(sf::Vector2f(30.f - ((carobj.health / carobj.maxHealth) * 30.f), 3.f));
 
@@ -302,30 +302,31 @@ void GameDisplay::drawGame()
 
 		// car background layer
 		car.setColor(carobj.getColor());
-        car.setTexture(this->getTexture("car/" + carobj.getTextureName()));
+        car.setTexture(getTexture("car/" + carobj.getTextureName()));
 		car.setTextureRect(IntRect(0, 0, 100, 40));
-        this->renderWnd->draw(car);
+        renderWnd->draw(car);
 
 		// car foreground layer (animated)
 		car.setColor(Color::White);
 		car.setTextureRect(IntRect(0, (animFrame + 1) * 40, 100, 40));
-		this->renderWnd->draw(car);
+		renderWnd->draw(car);
 
 		// reset
         car.setScale(2.0f, 2.0f);
         car.setRotation(0.f);
 
+        // ...health bars
         rect1.setPosition(car.getPosition().x, car.getPosition().y + 30);
         rect2.setPosition(car.getPosition().x + rect1.getSize().x, car.getPosition().y + 30);
 
         if(!game->cars[i]->isDestroying())
         {
-            this->renderWnd->draw(rect1);
-            this->renderWnd->draw(rect2);
+            renderWnd->draw(rect1);
+            renderWnd->draw(rect2);
         }
     }
 
-    this->drawEffect();
+    drawEffect();
 }
 
 void GameDisplay::drawEffect()
@@ -335,9 +336,9 @@ void GameDisplay::drawEffect()
     if(!game->paused())
     {
         if(game->powerTime > 1)
-            game->powerRegistry.find(game->getCurrentPower())->second->drawPower(this->renderWnd);
+            game->powerRegistry.find(game->getCurrentPower())->second->drawPower(renderWnd);
         else if(game->getCurrentPower() != -1)
-            game->powerRegistry.find(game->getCurrentPower())->second->drawPowerIdle(this->renderWnd);
+            game->powerRegistry.find(game->getCurrentPower())->second->drawPowerIdle(renderWnd);
     }
 }